+++ src/Console/Commands/Recount.php
@@ -15,13 +15,11 @@ namespace Cog\Laravel\Love\Console\Commands;
 
 use Cog\Contracts\Love\Reactable\Exceptions\ReactableInvalid;
 use Cog\Contracts\Love\Reactable\Models\Reactable as ReactableContract;
-use Cog\Contracts\Love\Reactant\Models\Reactant as ReactantContract;
+use Cog\Laravel\Love\Reactant\Jobs\RebuildReactionAggregatesJob;
 use Cog\Laravel\Love\Reactant\Models\Reactant;
-use Cog\Laravel\Love\Reactant\ReactionCounter\Models\ReactionCounter;
-use Cog\Laravel\Love\Reactant\ReactionCounter\Services\ReactionCounterService;
-use Cog\Laravel\Love\Reactant\ReactionTotal\Models\ReactionTotal;
 use Cog\Laravel\Love\ReactionType\Models\ReactionType;
 use Illuminate\Console\Command;
+use Illuminate\Contracts\Bus\Dispatcher;
 use Illuminate\Database\Eloquent\Relations\Relation;
 use Symfony\Component\Console\Input\InputOption;
 
@@ -42,14 +40,40 @@ final class Recount extends Command
     protected $description = 'Recount reactions of the reactable models';
 
     /**
+     * @var \Illuminate\Contracts\Bus\Dispatcher
+     */
+    private $dispatcher;
+
+    /**
+     * Get the console command options.
+     *
+     * @return array
+     */
+    protected function getOptions(
+    ): array {
+        return [
+            ['model', null, InputOption::VALUE_OPTIONAL, 'The name of the reactable model'],
+            ['type', null, InputOption::VALUE_OPTIONAL, 'The name of the reaction type'],
+            ['queue-connection', null, InputOption::VALUE_OPTIONAL, 'The name of the queue connection'],
+        ];
+    }
+
+    public function __construct(
+        Dispatcher $dispatcher
+    ) {
+        parent::__construct();
+        $this->dispatcher = $dispatcher;
+    }
+
+    /**
      * Execute the console command.
      *
      * @return void
      *
      * @throws \Cog\Contracts\Love\Reactable\Exceptions\ReactableInvalid
      */
-    public function handle(): void
-    {
+    public function handle(
+    ): void {
         if ($reactableType = $this->option('model')) {
             $reactableType = $this->normalizeReactableModelType($reactableType);
         }
@@ -58,50 +82,24 @@ final class Recount extends Command
             $reactionType = ReactionType::fromName($reactionType);
         }
 
-        $reactantsQuery = Reactant::query();
-
-        if ($reactableType) {
-            $reactantsQuery->where('type', $reactableType);
+        $queueConnectionName = $this->option('queue-connection');
+        if ($queueConnectionName === null || $queueConnectionName === '') {
+            $queueConnectionName = 'sync';
         }
 
-        $reactants = $reactantsQuery->get();
+        $reactants = $this->collectReactants($reactableType);
+
+        $this->warnProcessingStartedOn($queueConnectionName);
         $this->getOutput()->progressStart($reactants->count());
         foreach ($reactants as $reactant) {
-            /** @var \Illuminate\Database\Eloquent\Builder $query */
-            $query = $reactant->reactions();
+            $this->dispatcher->dispatch(
+                (new RebuildReactionAggregatesJob($reactant, $reactionType))
+                    ->onConnection($queueConnectionName)
+            );
 
-            if ($reactionType) {
-                $query->where('reaction_type_id', $reactionType->getId());
-            }
-
-            $counters = $reactant->getReactionCounters();
-
-            /** @var \Cog\Laravel\Love\Reactant\ReactionCounter\Models\ReactionCounter $counter */
-            foreach ($counters as $counter) {
-                if ($reactionType && $counter->isNotReactionOfType($reactionType)) {
-                    continue;
-                }
-
-                $counter->update([
-                    'count' => ReactionCounter::COUNT_DEFAULT,
-                    'weight' => ReactionCounter::WEIGHT_DEFAULT,
-                ]);
-            }
-
-            $reactions = $query->get();
-            $this->recountCounters($reactant, $reactions);
-            $this->recountTotal($reactant);
             $this->getOutput()->progressAdvance();
         }
         $this->getOutput()->progressFinish();
-    }
-
-    protected function getOptions(): array
-    {
-        return [
-            ['model', null, InputOption::VALUE_OPTIONAL, 'The name of the reactable model'],
-            ['type', null, InputOption::VALUE_OPTIONAL, 'The name of the reaction type'],
-        ];
     }
 
     /**
@@ -135,7 +133,7 @@ final class Recount extends Command
             $modelType = $this->findModelTypeInMorphMap($modelType);
         }
 
-        $model = new $modelType;
+        $model = new $modelType();
 
         if (!$model instanceof ReactableContract) {
             throw ReactableInvalid::notImplementInterface($modelType);
@@ -164,39 +162,42 @@ final class Recount extends Command
         return $morphMap[$modelType];
     }
 
-    private function recountTotal(
-        ReactantContract $reactant
-    ): void {
-        $counters = $reactant->getReactionCounters();
+    /**
+     * Collect all reactants we want to affect.
+     *
+     * @param string|null $reactableType
+     * @return \Cog\Contracts\Love\Reactant\Models\Reactant[]|\Illuminate\Database\Eloquent\Collection
+     */
+    private function collectReactants(
+        ?string $reactableType = null
+    ): iterable {
+        $reactantsQuery = Reactant::query();
 
-        if (count($counters) === 0) {
-            return;
+        if ($reactableType !== null) {
+            $reactantsQuery->where('type', $reactableType);
         }
 
-        $totalCount = ReactionTotal::COUNT_DEFAULT;
-        $totalWeight = ReactionTotal::WEIGHT_DEFAULT;
-
-        foreach ($counters as $counter) {
-            $totalCount += $counter->getCount();
-            $totalWeight += $counter->getWeight();
-        }
-
-        /** @var \Cog\Laravel\Love\Reactant\ReactionTotal\Models\ReactionTotal $total */
-        $total = $reactant->getReactionTotal();
-        $total->update([
-            'count' => $totalCount,
-            'weight' => $totalWeight,
-        ]);
+        return $reactantsQuery->get();
     }
 
-    private function recountCounters(
-        ReactantContract $reactant,
-        iterable $reactions
+    /**
+     * Write warning output that processing has been started.
+     *
+     * @param string|null $queueConnectionName
+     * @return void
+     */
+    private function warnProcessingStartedOn(
+        ?string $queueConnectionName
     ): void {
-        $service = new ReactionCounterService($reactant);
-
-        foreach ($reactions as $reaction) {
-            $service->addReaction($reaction);
+        if ($queueConnectionName === 'sync') {
+            $message = 'Rebuilding reaction aggregates synchronously.';
+        } else {
+            $message = sprintf(
+                'Adding rebuild reaction aggregates to the `%s` queue connection.',
+                $queueConnectionName
+            );
         }
+
+        $this->warn($message);
     }
 }